Fix:Updating Conflict API

+++ entrypoint.sh
@@ -3,7 +3,7 @@ set -e
 
 cd /var/www/html
 
-if [ ! -f .env ]; then
+if [ ! -f .env ] && [ -f .env.example ]; then
   cp .env.example .env
 fi
